Hard Fork 1 - Remove Adaptive Block Sizes to Speed Up Blockchain Sync
Hard Fork 1
We were getting blocks slower than 9 seconds because of the adaptive block size calculations. To calculate the median block size, 576,000 block timestamps were needed. This is the number of blocks produced in 60 days to prevent attacks to artificially increasing the block size due to our 0 fees. This large number of timestamps caused getting blocks templates to be delayed and block submissions to be delayed, resulting in 15 second blocks at the fastest. We decided to remove the adpative block size feature to decrease the block synchronization times.