Added Check Payment
Check Payment

This commit adds the ability for a sender to prove that funds were sent to a recipient. The information required from the sender are the transaction id, transaction secret key, and the receiver's address. All three can be obtained by the sender using simplewallet and the command "outgoingTransactions". This feature is required by the Bisq exchange in the event that there is a dispute about funds not being received. If the sender says that he has already sent Cash2 but the receivers says that he did not receive any Cash2, the sender can prove that the money was sent by using the Cash2 block explorer to check the payment using the transaction id, transaction secret key, and the receiver's address.

The code commited here is heavily borrowed from the Karbo source code. I would like to give the biggest thanks to their development team because they have contribute imensely to the CryptoNote community. We are much in debted to you guys.