Hard Fork 2 - Fix Block Difficulty Integer Overflow
Problem - getblocktemplate fails because the nextDifficulty function returns false because the total work done overflows 64 bits or is greater than 18,446,744,073,709,551,615.

Cause - Using Blake2b as the hashing algorithm resulted in block difficulties being as high as 620 trillion with just 40 miners. Blocks keep record of a value called cumulative difficulties which is the sum of the difficulties of the last 3,600 blocks and is used to calculate a value called total work.  On block height 420,015, the total work was 2,049,660,151,821,555,415. The total work is then multiplied by 9 giving a product larger than 18 quadrillion, which is the maximum value of a uint64_t.

Fix - During mining, a hash of a block is calculated. For instance the hash of block 416,316 is 00000000000050431c02b02e79bcd3b0433e797cbc35551cd4e81a8eb10c59d1 and the network difficulty is 561,034,697,440,408. To figure out if this hash works with this difficulty, we take the hash and multiply it with the difficulty, and the product must be less than 2^256. The new changes that were put in place is to chop off the first 10 zeros of the hash. This is the same as multiplying the hash by 1,099,511,627,776. We also must divide the network difficulty by this number. So 561,034,697,440,408 / 1,099,511,627,776 = 510. So now the network difficulty is 510 and not 561,034,697,440,408 anymore. This allows us to keep the network difficulty value from being too large, but we must remember to multiply it by 1,099,511,627,776 to get the true network difficulty. So the new equation is 2^256 = reducedNetworkDifficulty x hash x 1,099,511,627,776. We reduce the difficulty by a factor of 16^10, so we must also increase the hash by a factor of 16^10 by chopping off the first 10 zeros of the hash.