Removed code for hard fork 1
Hardfork 1 occurred at block height 230,500. The purpose of this hardfork was to discontinue calculating the block reward based on the median block size. Instead, the block reward is calculated based on the block's size. If the block's size is less than a calculated amount, the block reward is awarded in full. However, all blocks prior to height 230,500 had block sizes that were less than this calculated amount, and therefore, there were no blocks that broke the new consensus rules.