Added tx fee of 0.01, limit tx extra size, check base tx signatures empty
This commit is to help prevent blockchain spamming. This consensus change is a soft fork and will be implemented starting at block height 1,100,000. The minimum fee is increased from 0 to 0.01, the transaction extra size is limited to 256 bytes, and all coinbase transactions are checked to make sure they do not have any signatures. Thank you to the Turtlecoin developers for recognizing these issues and for providing fixes for them. Thank you to the Karbo developers for also implementing fixes for this issue and also for their previous work and guidance on how to increase the minimal transaction fee.

Hard fork height 1 was originally set to height 230,500 but is now removed from the code because all blocks prior to height 230,500 follows the new consensus rules implemented. Therefore, a hard fork did not really occur at height 230,500.