Trying to fix remote node error "The stream is ended"
Any node that runs with the flag --rpc-bind-ip=0.0.0.0 has issues dropping connections to anyone who is connected to it. We believe this issue is caused by the failure to bind the accept loop after an exception is thrown in src/Rpc/HttpServer.cpp as pointed out by many other CryotNote projects. The solution we are using here was borrowed from the Karbo project, and we would like to give a big thanks to them.