Change mined block unlock window from 60 to 400
Hardfork at block height 1,700,000 to change CRYPTONOTE_MINED_MONEY_UNLOCK_WINDOW from 60 to 400. This will require miners to wait for at least 400 blocks or about 1 hour before they can spend a coinbase transaction from a block they mined. The purpose of this change is to prepare Cash2 to be listed on exchanges. We feel that it is safer for exchanges if miners had to wait a longer period of time before they are able to deposit their funds on to an exchange due to the posibility of chain reorganizations and those funds being invalid.

Also, the dust threshold has been reduced from 10000000 to 0. This increased the number of outputs in the coinbase transaction from 3 outputs to 8 outputs.

There is also a fix made to the constructMinerTx() functions in Currency.cpp to calculate the dust threshold based on the block height, which was overlooked in the last software upgrade in version 4.2.