Added checkpoints to prevent chain reorganization
Added checkpoints and other measures to prevent an alternative block from being added to the chain prior to HARD_FORK_HEIGHT_2 which reset the cummulative difficulties of the blocks. Without this security measure, an alternative block that is added to the chain prior to  HARD_FORK_HEIGHT_2 will cause the blockchain to reorganize to make this new alternative block the tip of the blockchain becuase its cummulative difficulty is larger than the current main chain's cummulative difficulty.