Walletd RPC : Added password for RPC request authentication
Walletd JSON RPC requests did not require authentication and had the potential for an attacker to gain control of the wallet by sending JSON RPC commands to the wallet via CSRF. Users can now run the Wallet JSON RPC server with the flag --rpc-password to require that all RPC commands include a password.

Thank you to the Turtlecoin developers for their work. Much of this code is borrowed from their commit turtlecoin/turtlecoin@4949e91

More information about this vulnerability can be found here -
https://www.cvedetails.com/cve/CVE-2018-1000093/
cryptonotefoundation#172

This commit is a copyof 2 commits from the master branch of Cash2
https://github.com/cash2/cash2/commit/1b9691aba632f5b69f39c749a64dfbfb57ffb33f
https://github.com/cash2/cash2/commit/f01b904c0736036dea5adb00b7a6314176d0f9f6