Walletd : Tells if the wallet is connected to the daemon
When starting Walletd, this new feature will tell the user whether the wallet is able to connect to the remote daemon or not. Prior to this, there was no indication or warning if the wallet was not connected to the daemon.

Reason for changes : To alert the user if there is an error connecting the wallet to the remote daemon. SouthXchange has requested this feature to be added.