Walletd : Changed RPC command create_address() to save the wallet file and to return the spend private key
When the RPC command create_address() is called, a new address is created, the wallet file is saved, the addresses is returned, and the spend private key for the address is returned. This ensures that everytime a new address is generated that the wallet file is saved and the spend private key is returned to the caller.

Reason for changes : to better protect against lost spend private keys when a new address is created.

This commit is a close duplicate of commit 208d1f372decbea7f75f970e95762e41cd1ec2e1 in the master branch https://github.com/cash2/cash2/commit/208d1f372decbea7f75f970e95762e41cd1ec2e1