Walletd : Automatically save wallet when an address is generated using a spend private key
When functions createAddress() with a specified spend private key and createAddresses() are called, the spend private key and address are automatically saved to disk.

Reason for changes : automatically save spend private keys to the wallet file to prevent accidental loss of keys.

This commit is a close duplicate of commit f803a069c1fded581ac38b5ecc4b31f326616454 from the master branch