Walletd : Changed default RPC connection port to the daemon to 12276
Reason for changes : The original port specified 8081 is not currently set to anything for the Cash2 project. Changed the defalut daemon port for Walletd to 12276 which is the default daemon port for all Cash2 nodes.

This commit is a duplicate of commit  72f49001a5962161611ab5db180b54c0faaf1c1b from the master branch