Walletd : Added command getSpendPrivateKeys()
Returns the spend private keys of all the addresses in your Walletd container.

Reason for changes : to be able to easily get all spend private keys of the wallet container to backup the wallet container.

This commit is a close duplicate of commit 4d50526602e3d6bfffda2f9780aeda999ba6cc22 from the master branch.