Walletd : Fixed formatting when printing block height during blockchain synchronization
The Walletd program shows the block height as the wallet continues to synchronize with the blockchain. When the wallet is reset, the formatting of the block height print out was messed up and had additional zeros added on the end. I researched online to find that the best way to fix this formatting issue was to set a fixed width for what is being printed out. The fixed with for the block heigt is set to 9 digits because currently the max block height for the Cash2 network is 500,000,000.

Reason for changes : fix formatting issues