Unit Tests : Fixed WalletApi.transferSendsChangeToAddress test from failing
Reason for changes : Originally the value fro MONEY was lower than the transaction fee which is 10000000. Therefore, the transaction would fail because the amount sent was not enough. Increasing the amount to more than the transaction fee allows this test to pass.

This commit is a duplicate of commit 9c7e5337bea0de5115b75348a9d529314ea04e2a from the master branch