Unit Tests : Fixed WalletLegacyApi.checkChange test from failing
eason for changes : Originally the value for banknote was equal to the transaction fee which is 10000000. Therefore, the transaction would fail because the amount sent plus the fee was greater than the banknote variable. Increasing the banknote variable to more than the transaction fee plus the sentAmount allows this test to pass.

This commit is a duplicate of commit 3ce6a22ee06cc03f55e6678312416fc9abecd74d from the master branch.