Core Tests : Fixed multiple core tests to make them runnable
Reason for changes : The transaction amounts that were being created in the core test code were too large because Cash2 has a block reward of 0.894069671 or less, so the accounts did not have enough money after mining on a few blocks to send transactions with such high amounts. This caused the tests to fail.

This commit is a duplicate of commit 148c85b4b7b61a2c1c3cdd6468deff6e1a76ebb0 in the master branch