Walletd : RPC commands createAddress() and createAddress(spendPrivateKey) no longer return the spend private key of the address created
Reason for changes : Previously in commit 6273b5cadf3d388c9c65c475561ab83395b944b9 from the exchanges branch and commit 208d1f372decbea7f75f970e95762e41cd1ec2e1 in the master branch, a change was made where the createAddress() RPC commands would return the spend private key of the newly created address. This was put in place to make it easier to the caller to keep record of the new spend private key. I was asked by the SouthXchange team to remove this feature as they felt that passing private keys outside of the wallet file is a security risk.